new file:   backend/db/src/comparisons.rs 	modified:   backend/db/src/filters.rs 	modified:   backend/db/src/lib.rs 	modified:   backend/db/src/models/announcements.rs 	modified:   backend/db/src/models/assignment_interpreter.rs 	modified:   backend/db/src/models/module.rs 	modified:   backend/db/src/models/plagiarism_case.rs 	modified:   backend/db/src/models/ticket_messages.rs 	modified:   backend/db/src/models/tickets.rs 	new file:   backend/db/src/repositories/announcement_repository.rs 	modified:   backend/db/src/repositories/assignment_file_repository.rs 	new file:   backend/db/src/repositories/assignment_interpreter_repository.rs 	modified:   backend/db/src/repositories/assignment_memo_output_repository.rs 	modified:   backend/db/src/repositories/assignment_overwrite_file_repository.rs 	modified:   backend/db/src/repositories/assignment_repository.rs 	modified:   backend/db/src/repositories/assignment_submission_output_repository.rs 	modified:   backend/db/src/repositories/assignment_submission_repository.rs 	modified:   backend/db/src/repositories/assignment_task_repository.rs 	modified:   backend/db/src/repositories/mod.rs 	modified:   backend/db/src/repositories/module_repository.rs 	modified:   backend/db/src/repositories/password_reset_token_repository.rs 	modified:   backend/db/src/repositories/plagiarism_case_repository.rs 	modified:   backend/db/src/repositories/ticket_message_repository.rs 	modified:   backend/db/src/repositories/ticket_repository.rs 	modified:   backend/db/src/repositories/user_module_role_repository.rs 	modified:   backend/db/src/repositories/user_repository.rs 	new file:   backend/services/src/announcement_service.rs 	modified:   backend/services/src/assignment_file_service.rs 	new file:   backend/services/src/assignment_interpreter_service.rs 	modified:   backend/services/src/assignment_memo_output_service.rs 	modified:   backend/services/src/assignment_overwrite_file_service.rs 	modified:   backend/services/src/assignment_service.rs 	modified:   backend/services/src/assignment_submission_service.rs 	modified:   backend/services/src/assignment_task_service.rs 	modified:   backend/services/src/lib.rs 	modified:   backend/services/src/module_service.rs 	modified:   backend/services/src/password_reset_token_service.rs 	modified:   backend/services/src/plagiarism_case_service.rs 	modified:   backend/services/src/ticket_message_service.rs 	modified:   backend/services/src/ticket_service.rs 	modified:   backend/services/src/user_module_role_service.rs 	modified:   backend/services/src/user_service.rs

+++ backend/db/src/models/module.rs
@@ -8,9 +8,9 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
     pub code: String,
-    pub year: i32,
+    pub year: i64,
     pub description: Option<String>,
-    pub credits: i32,
+    pub credits: i64,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
